Updates Boid class
* Moves simplified Boid class into boid.py
* Adds Boid import to sim
